+++ Mouna.Api.Crud/BusinessLogic/Interfaces/IEmployeeService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Mouna.Api.Crud.BusinessLogic.Interfaces
+{
+    public interface IEmployeeService
+    {
+    }
+}
