+++ Mouna.Api.Crud/BusinessLogic/Models/EmployeeBLL.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Mouna.Api.Crud.BusinessLogic.Models
+{
+    public class EmployeeBLL
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Salary { get; set; }
+    }
+}
