+++ Mouna.Api.Crud/BusinessLogic/Services/EmployeeService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Mouna.Api.Crud.BusinessLogic.Services
+{
+    public class EmployeeService
+    {
+    }
+}
